djangogram/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.33.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
decorator==5.1.1
defusedxml==0.7.1
Django==4.2.1
django-bootstrap4==23.1
django-debug-toolbar==4.0.0
django-environ==0.10.0
django-extensions==3.2.1
django-filter==23.2
django-model-utils==4.3.1
django-rest-swagger==2.2.0
django-storages==1.13.2
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-jwt-5==1.13.0
easy-thumbnails==2.8.5
exceptiongroup==1.1.1
Faker==11.0.0
freetype-py==2.3.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.3
MarkupSafe==2.1.2
mixer==7.2.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
pip==23.1.2
pluggy==1.0.0
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycairo==1.22.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.0.0
pytest==7.3.1
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
# reportlab==4.0.0
requests==2.31.0
requests-oauthlib==1.3.1
rlPyCairo==0.2.0
rsa==4.9
self==2020.12.3
simplejson==3.19.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
social-auth-core==4.2.0
soupsieve==2.4.1
sqlparse==0.4.4
tomli==2.0.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.5
zipp==3.15.0
